move processBlocks to full sync it will be a syncronizer based behavior
