+++ packages/client/lib/sync/fullsync.ts
@@ -1,9 +1,20 @@
+import { Hardfork } from '@ethereumjs/common'
 import { BN } from 'ethereumjs-util'
 import { Peer } from '../net/peer/peer'
 import { short } from '../util'
+import { Event } from '../types'
 import { Synchronizer, SynchronizerOptions } from './sync'
 import { BlockFetcher } from './fetcher'
-import type { Block } from '@ethereumjs/block'
+import { VMExecution } from '../execution'
+import type { Block, BlockHeader } from '@ethereumjs/block'
+import type { TxPool } from '../service/txpool'
+
+interface FullSynchronizerOptions extends SynchronizerOptions {
+  /** Tx Pool */
+  txPool: TxPool
+  execution: VMExecution
+}
+
 interface HandledObject {
   hash: Buffer
   added: number
@@ -15,11 +26,19 @@ type PeerId = string
  * @memberof module:sync
  */
 export class FullSynchronizer extends Synchronizer {
+  private txPool: TxPool
+  private execution: VMExecution
   private newBlocksKnownByPeer: Map<PeerId, HandledObject[]>
 
-  constructor(options: SynchronizerOptions) {
+  constructor(options: FullSynchronizerOptions) {
     super(options)
+    this.txPool = options.txPool
+    this.execution = options.execution
     this.newBlocksKnownByPeer = new Map()
+
+    this.config.events.on(Event.SYNC_FETCHER_FETCHED, this.processBlocks)
+    this.config.events.on(Event.SYNC_EXECUTION_VM_ERROR, this.stop)
+
     void this.chain.update()
   }
 
@@ -42,6 +61,7 @@ export class FullSynchronizer extends Synchronizer {
     const hash = this.chain.blocks.latest!.hash()
     this.startingBlock = number
     this.config.chainCommon.setHardforkByBlockNumber(number, td)
+
     this.config.logger.info(
       `Latest local block number=${Number(number)} td=${td} hash=${short(
         hash
@@ -247,11 +267,80 @@ export class FullSynchronizer extends Synchronizer {
     }
   }
 
+  async stop(): Promise<boolean> {
+    this.config.events.on(Event.SYNC_FETCHER_FETCHED, this.processBlocks)
+    this.config.events.on(Event.SYNC_EXECUTION_VM_ERROR, this.stop)
+
+    return await super.stop()
+  }
+
   /**
    * Close synchronizer.
    */
   async close() {
     if (!this.opened) return
     await super.close()
+  }
+
+  async processBlocks(blocks: Block[] | BlockHeader[]) {
+    if (this.config.chainCommon.gteHardfork(Hardfork.Merge)) {
+      if (this.fetcher !== null) {
+        // If we are beyond the merge block we should stop the fetcher
+        this.config.logger.info('Merge hardfork reached, stopping block fetcher')
+        this.clearFetcher()
+      }
+    }
+
+    if (blocks.length === 0) {
+      if (this.fetcher !== null) {
+        this.config.logger.warn('No blocks fetched are applicable for import')
+      }
+      return
+    }
+
+    blocks = blocks as Block[]
+    const first = blocks[0].header.number
+    const last = blocks[blocks.length - 1].header.number
+    const hash = short(blocks[0].hash())
+    const baseFeeAdd = this.config.chainCommon.gteHardfork(Hardfork.London)
+      ? `baseFee=${blocks[0].header.baseFeePerGas} `
+      : ''
+
+    let attentionHF: string | null = null
+    const nextHFBlockNum = this.config.chainCommon.nextHardforkBlockBN()
+    if (nextHFBlockNum !== null) {
+      const remaining = nextHFBlockNum.sub(last)
+      if (remaining.lten(10000)) {
+        const nextHF = this.config.chainCommon.getHardforkByBlockNumber(nextHFBlockNum)
+        attentionHF = `${nextHF} HF in ${remaining} blocks`
+      }
+    } else {
+      if (
+        this.config.chainCommon.hardfork() === Hardfork.MergeForkIdTransition &&
+        !this.config.chainCommon.gteHardfork(Hardfork.Merge)
+      ) {
+        const mergeTD = this.config.chainCommon.hardforkTD(Hardfork.Merge)!
+        const td = this.chain.blocks.td
+        const remaining = mergeTD.sub(td)
+        if (remaining.lte(mergeTD.divn(10))) {
+          attentionHF = `Merge HF in ${remaining} TD`
+        }
+      }
+    }
+
+    this.config.logger.info(
+      `Imported blocks count=${
+        blocks.length
+      } first=${first} last=${last} hash=${hash} ${baseFeeAdd}hardfork=${this.config.chainCommon.hardfork()} peers=${
+        this.pool.size
+      }`,
+      { attentionHF }
+    )
+
+    this.txPool.removeNewBlockTxs(blocks)
+
+    if (!this.running) return
+    await this.execution.run()
+    this.txPool.checkRunState()
   }
 }
