+++ packages/client/test/sync/fullsync.spec.ts
@@ -7,6 +7,8 @@ import { Event } from '../../lib/types'
 import { Block } from '@ethereumjs/block'
 
 tape('[FullSynchronizer]', async (t) => {
+  const txPool: any = { removeNewBlockTxs: () => {}, checkRunState: () => {} }
+  const execution: any = { run: () => {} }
   class PeerPool {
     open() {}
     close() {}
@@ -32,7 +34,7 @@ tape('[FullSynchronizer]', async (t) => {
     const config = new Config({ transports: [] })
     const pool = new PeerPool() as any
     const chain = new Chain({ config })
-    const sync = new FullSynchronizer({ config, pool, chain })
+    const sync = new FullSynchronizer({ config, pool, chain, txPool, execution })
     t.equals(sync.type, 'full', 'full type')
     t.end()
   })
@@ -45,6 +47,8 @@ tape('[FullSynchronizer]', async (t) => {
       config,
       pool,
       chain,
+      txPool,
+      execution,
     })
     ;(sync as any).pool.open = td.func<PeerPool['open']>()
     ;(sync as any).pool.peers = []
@@ -59,7 +63,7 @@ tape('[FullSynchronizer]', async (t) => {
     const config = new Config({ transports: [] })
     const pool = new PeerPool() as any
     const chain = new Chain({ config })
-    const sync = new FullSynchronizer({ config, pool, chain })
+    const sync = new FullSynchronizer({ config, pool, chain, txPool, execution })
     const peer = { eth: { getBlockHeaders: td.func(), status: { bestHash: 'hash' } } }
     const headers = [{ number: new BN(5) }]
     td.when(peer.eth.getBlockHeaders({ block: 'hash', max: 1 })).thenResolve([new BN(1), headers])
@@ -79,6 +83,8 @@ tape('[FullSynchronizer]', async (t) => {
       interval: 1,
       pool,
       chain,
+      txPool,
+      execution,
     })
     ;(sync as any).running = true
     ;(sync as any).height = td.func()
@@ -111,6 +117,8 @@ tape('[FullSynchronizer]', async (t) => {
       interval: 1,
       pool,
       chain,
+      txPool,
+      execution,
     })
     sync.best = td.func<typeof sync['best']>()
     sync.latest = td.func<typeof sync['latest']>()
@@ -149,6 +157,8 @@ tape('[FullSynchronizer]', async (t) => {
       interval: 1,
       pool,
       chain,
+      txPool,
+      execution,
     })
     ;(sync as any).fetcher = {
       enqueueByNumberList: (blockNumberList: BN[], min: BN) => {
